requirements.txt
# Celery和Redis相关
celery==5.3.4
redis==5.0.1
kombu==5.3.4
flower==2.0.1

# Web框架
flask==3.0.0
flask-cors==4.0.0
flask-restful==0.3.10
requests

# 数据处理
pydantic==2.5.0
python-dateutil==2.8.2
croniter==2.0.1

# 日志和监控
structlog==23.2.0
prometheus-client==0.19.0

# 安全和认证
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2

# 开发和测试
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0

# 生产环境
gunicorn==21.2.0
supervisor==4.2.5 